+++ docs/monitors/collectd-kafka_consumer.md
@@ -1,78 +1,28 @@
 <!--- GENERATED BY gomplate from scripts/docs/monitor-page.md.tmpl --->
 
-# collectd/hadoopjmx
+# collectd/kafka_consumer
 
- Collects metrics about a Hadoop cluster using
-using collectd's GenericJMX plugin.
+ Monitors a java based Kafka consumer using GenericJMX.
 
-Also see
-https://github.com/signalfx/integrations/tree/master/collectd-hadoop.
+See the [integration documentation](https://github.com/signalfx/integrations/tree/master/collectd-kafka_consumer)
+for more information.
 
->To enable JMX in Hadoop, add the following JVM options to hadoop-env.sh and yarn-env.sh respectively
-
-**hadoop-env.sh:**
-```
-export HADOOP_NAMENODE_OPTS="-Dcom.sun.management.jmxremote.ssl=false -Dcom.sun.management.jmxremote.authenticate=false -Dcom.sun.management.jmxremote.port=5677 $HADOOP_NAMENODE_OPTS"
-export HADOOP_DATANODE_OPTS="-Dcom.sun.management.jmxremote.ssl=false -Dcom.sun.management.jmxremote.authenticate=false -Dcom.sun.management.jmxremote.port=5679 $HADOOP_DATANODE_OPTS"
-```
-
-**yarn-env.sh:**
-```
-export YARN_NODEMANAGER_OPTS="-Dcom.sun.management.jmxremote.ssl=false -Dcom.sun.management.jmxremote.authenticate=false -Dcom.sun.management.jmxremote.port=5678 $YARN_NODEMANAGER_OPTS"
-export YARN_NODEMANAGER_OPTS="-Dcom.sun.management.jmxremote.ssl=false -Dcom.sun.management.jmxremote.authenticate=false -Dcom.sun.management.jmxremote.port=5678 $YARN_NODEMANAGER_OPTS"
-```
-
-This monitor has a set of built in MBeans configured for:
-- [Name Nodes](https://github.com/signalfx/signalfx-agent/tree/master/internal/monitors/collectd/hadoopjmx/nameNodeMBeans.go)
-- [Resource Manager](https://github.com/signalfx/signalfx-agent/tree/master/internal/monitors/collectd/hadoopjmx/resourceManagerMBeans.go)
-- [Node Manager](https://github.com/signalfx/signalfx-agent/tree/master/internal/monitors/collectd/hadoopjmx/nodeManagerMBeans.go)
-- [Data Nodes](https://github.com/signalfx/signalfx-agent/tree/master/internal/monitors/collectd/hadoopjmx/dataNodeMBeans.go)
+This monitor has a set of [built in MBeans
+configured](https://github.com/signalfx/signalfx-agent/tree/master/internal/monitors/collectd/kafka_consumer/mbeans.go)
+for which it pulls metrics from the Kafka consumer's JMX endpoint.
 
 Sample YAML configuration:
-
-	Name Node
 ```yaml
 monitors:
-- type: collectd/hadoopjmx
-  host: 127.0.0.1
-  port: 5677
-  nodeType: nameNode
+  - type: collectd/kafka_consumer
+    host: localhost
+    port: 9099
 ```
 
-	Resource Manager
-```yaml
-monitors:
-- type: collectd/hadoopjmx
-  host: 127.0.0.1
-  port: 5680
-  nodeType: resourceManager
-```
 
-	Node Manager
-```yaml
-monitors:
-- type: collectd/hadoopjmx
-  host: 127.0.0.1
-  port: 8002
-  nodeType: nodeManager
-```
+Monitor Type: `collectd/kafka_consumer`
 
-	Data Node
-```yaml
-monitors:
-- type: collectd/hadoopjmx
-  host: 127.0.0.1
-  port: 5679
-  nodeType: dataNode
-```
-
-You may also configure the [collectd/hadoop](https://github.com/signalfx/signalfx-agent/tree/master/docs/monitors/collectd/hadoop)
-monitor to collect additional metrics about the hadoop cluster from the REST API
-
-
-Monitor Type: `collectd/hadoopjmx`
-
-[Monitor Source Code](https://github.com/signalfx/signalfx-agent/tree/master/internal/monitors/collectd/hadoopjmx)
+[Monitor Source Code](https://github.com/signalfx/signalfx-agent/tree/master/internal/monitors/collectd/kafkaconsumer)
 
 **Accepts Endpoints**: **Yes**
 
@@ -94,7 +44,6 @@ Monitor Type: `collectd/hadoopjmx`
 | `mBeansToCollect` | no | `list of string` | A list of the MBeans defined in `mBeanDefinitions` to actually collect. If not provided, then all defined MBeans will be collected. |
 | `mBeansToOmit` | no | `list of string` | A list of the MBeans to omit. This will come handy in cases where only a few MBeans need to omitted from the default list |
 | `mBeanDefinitions` | no | `map of object (see below)` | Specifies how to map JMX MBean values to metrics.  If using a specific service monitor such as cassandra, kafka, or activemq, they come pre-loaded with a set of mappings, and any that you add in this option will be merged with those.  See [collectd GenericJMX](https://collectd.org/documentation/manpages/collectd-java.5.shtml#genericjmx_plugin) for more details. |
-| `nodeType` | **yes** | `string` | Hadoop Node Type |
 
 
 The **nested** `mBeanDefinitions` config object has the following fields:
