--- internal/monitors/collectd/kafka_producer/kafka_producer.go
@@ -1,33 +1,0 @@
-// +build !windows
-
-package kafka_producer
-
-import (
-	"github.com/signalfx/signalfx-agent/internal/monitors"
-	"github.com/signalfx/signalfx-agent/internal/monitors/collectd/genericjmx"
-	yaml "gopkg.in/yaml.v2"
-)
-
-const monitorType = "collectd/kafka_producer"
-
-var serviceName = "kafka_producer"
-
-// Monitor is the main type that represents the monitor
-type Monitor struct {
-	*genericjmx.JMXMonitorCore
-}
-
-func init() {
-	var defaultMBeans genericjmx.MBeanMap
-	err := yaml.Unmarshal([]byte(defaultMBeanYAML), &defaultMBeans)
-	if err != nil {
-		panic("YAML for GenericJMX MBeans is invalid: " + err.Error())
-	}
-	defaultMBeans = defaultMBeans.MergeWith(genericjmx.DefaultMBeans)
-
-	monitors.Register(monitorType, func() interface{} {
-		return &Monitor{
-			genericjmx.NewJMXMonitorCore(defaultMBeans, serviceName),
-		}
-	}, &genericjmx.Config{})
-}
