+++ internal/monitors/collectd/kafkaconsumer/mbeans.go
@@ -1,4 +1,4 @@
-package kafka_consumer
+package kafkaconsumer
 
 var defaultMBeanYAML = `
 records-lag-max:
