+++ internal/monitors/collectd/kafkaproducer/mbeans.go
@@ -1,4 +1,4 @@
-package kafka_producer
+package kafkaproducer
 
 var defaultMBeanYAML = `
 response-rate:
@@ -66,64 +66,4 @@ io-wait-time-ns-avg:
     type: "gauge"
     table: false
     attribute: "io-wait-time-ns-avg"
-
-byte-rate-per-topic:
-  objectName: "kafka.producer:client-id=*,topic=*,type=producer-topic-metrics"
-  instancePrefix: "all"
-  dimensions:
-  - client-id
-  - topic
-  values:
-  - instancePrefix: "kafka.producer.byte-rate"
-    type: "gauge"
-    table: false
-    attribute: "byte-rate"
-
-compression-rate-per-topic:
-  objectName: "kafka.producer:client-id=*,topic=*,type=producer-topic-metrics"
-  instancePrefix: "all"
-  dimensions:
-  - client-id
-  - topic
-  values:
-  - instancePrefix: "kafka.producer.compression-rate"
-    type: "gauge"
-    table: false
-    attribute: "compression-rate"
-
-record-error-rate-per-topic:
-  objectName: "kafka.producer:client-id=*,topic=*,type=producer-topic-metrics"
-  instancePrefix: "all"
-  dimensions:
-  - client-id
-  - topic
-  values:
-  - instancePrefix: "kafka.producer.record-error-rate"
-    type: "gauge"
-    table: false
-    attribute: "record-error-rate"
-
-record-retry-rate-per-topic:
-  objectName: "kafka.producer:client-id=*,topic=*,type=producer-topic-metrics"
-  instancePrefix: "all"
-  dimensions:
-  - client-id
-  - topic
-  values:
-  - instancePrefix: "kafka.producer.record-retry-rate"
-    type: "gauge"
-    table: false
-    attribute: "record-retry-rate"
-
-record-send-rate-per-topic:
-  objectName: "kafka.producer:client-id=*,topic=*,type=producer-topic-metrics"
-  instancePrefix: "all"
-  dimensions:
-  - client-id
-  - topic
-  values:
-  - instancePrefix: "kafka.producer.record-send-rate"
-    type: "gauge"
-    table: false
-    attribute: "record-send-rate"
 `
